# documentation/api-docs.json
{"openapi":"3.1.0","info":{"title":"User Service E-Commerce API","description":"API documentation for E-Commerce Users microservices","version":"1.0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"security":[{"Bearer Authentication":[]}],"tags":[{"name":"Admin Management","description":"Admin operations for user management and system administration"},{"name":"User Management","description":"Operations for regular user management and profile access"},{"name":"Authentication","description":"Authentication operations including login, register, refresh token and logout"}],"paths":{"/admin/userupdate/{userid}":{"put":{"tags":["Admin Management"],"summary":"Update User","description":"Updates user details (admin only)","operationId":"updateUser","parameters":[{"name":"userid","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateRequest"}}},"required":true},"responses":{"200":{"description":"User updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"403":{"description":"Forbidden - requires admin role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/auth/user/register":{"post":{"tags":["Authentication"],"summary":"Register User","description":"Registers a new User","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterRequest"}}},"required":true},"responses":{"201":{"description":"User successfully registered","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"400":{"description":"Invalid input or user already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/user/login":{"post":{"tags":["Authentication"],"summary":"User Login","description":"Authenticates a user and returns JWT token","operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequest"}}},"required":true},"responses":{"200":{"description":"Authentication successful","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"401":{"description":"Invalid credentials","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/refresh-token":{"post":{"tags":["Authentication"],"summary":"Refresh access token","description":"Generates a new access token using the refresh token","operationId":"refreshToken","parameters":[{"name":"refreshToken","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Token refreshed successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"401":{"description":"Invalid or expired refresh token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout user","description":"Invalidates the current access token by adding it to the blacklist","operationId":"logout","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Logged out successfully","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Invalid or missing token","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/auth/logout-all":{"post":{"tags":["Authentication"],"summary":"Logout from all devices","description":"Invalidates all sessions by blacklisting all tokens","operationId":"logoutAll","parameters":[{"name":"refreshToken","in":"cookie","required":false,"schema":{"type":"string"}},{"name":"Refresh-Token","in":"header","required":false,"schema":{"type":"string"}},{"name":"refreshToken","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Logged out from all devices successfully","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Invalid or missing token","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/users/user/{id}":{"get":{"tags":["User Management"],"summary":"Get user by ID","description":"Fetch a user using their ID","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User details retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"403":{"description":"Forbidden - requires user role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"security":[{"Bearer Authentication":[]}]},"patch":{"tags":["User Management"],"summary":"Edit User Details","description":"Edit user details (user only)","operationId":"editUserDetails","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSelfEditRequest"}}},"required":true},"responses":{"200":{"description":"User details updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"403":{"description":"Forbidden - requires user role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/users/profile/{userId}":{"get":{"tags":["User Management"],"summary":"Get User Profile","description":"Retrieves user profile (own or admin access only)","operationId":"getUserProfile","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Profile retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"403":{"description":"Forbidden - can only view own profile or requires admin role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"security":[{"Bearer Authentication":[]}]}},"/admin/user/search":{"get":{"tags":["Admin Management"],"summary":"Search Users","description":"Searches for users by email or username (admin only)","operationId":"searchUser","parameters":[{"name":"email","in":"query","required":false,"schema":{"type":"string"}},{"name":"username","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Search results retrieved successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}},"403":{"description":"Forbidden - requires admin role","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}},"security":[{"Bearer Authentication":[]}]}},"/admin/user/all":{"get":{"tags":["Admin Management"],"summary":"Get All Users","description":"Fetches a list of all users (admin only)","operationId":"getAllUsers","responses":{"200":{"description":"List of all users retrieved successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}},"403":{"description":"Forbidden - requires admin role","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}},"security":[{"Bearer Authentication":[]}]},"delete":{"tags":["Admin Management"],"summary":"Delete All Users","description":"Deletes all users from the system except the current admin","operationId":"deleteAllUsers","responses":{"200":{"description":"All users deleted successfully except current admin","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden - requires admin role","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"Bearer Authentication":[]}]}},"/admin/dashboard":{"get":{"tags":["Admin Management"],"summary":"Admin Dashboard","description":"Displays welcome message for admin dashboard","operationId":"adminDashboard","responses":{"200":{"description":"Welcome message displayed","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Forbidden - requires admin role","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/users/users/{email}":{"delete":{"tags":["User Management"],"summary":"Delete User","description":"Deletes a user by their email (admin or self-deletion)","operationId":"deleteUser","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User deleted successfully","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"type":"object"}}}},"403":{"description":"Forbidden - can only delete own account or requires admin role","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"Bearer Authentication":[]}]}}},"components":{"schemas":{"Address":{"type":"object","properties":{"id":{"type":"string"},"street":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zipCode":{"type":"string"}}},"UserUpdateRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"email":{"type":"string"},"password":{"type":"string","minLength":1},"addresses":{"type":"array","items":{"$ref":"#/components/schemas/Address"}},"orderIds":{"type":"array","items":{"type":"string"}},"roles":{"type":"array","items":{"type":"string"}}}},"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string","minLength":1},"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1},"addresses":{"type":"array","items":{"$ref":"#/components/schemas/Address"}},"orderIds":{"type":"array","items":{"type":"string"}},"roles":{"type":"array","items":{"type":"string"}}}},"UserRegisterRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1},"name":{"type":"string","minLength":1},"address":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"msg":{"type":"string"}}},"UserLoginRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}}},"UserSelfEditRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1},"addresses":{"type":"array","items":{"$ref":"#/components/schemas/Address"}}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","name":"Bearer Authentication","scheme":"Bearer","bearerFormat":"JWT"}}}}