# src/main/java/genaidemopoc/ecommerceproj1a/jwtspringsecurity/usersvc/dto/UserLoginRequest.java
package genaidemopoc.ecommerceproj1a.jwtspringsecurity.usersvc.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;
import lombok.experimental.SuperBuilder;

@Data
@SuperBuilder
@NoArgsConstructor
@EqualsAndHashCode(callSuper = true)
public class UserLoginRequest extends BaseUserRequest {

	public UserLoginRequest(
			@NotBlank(message = "Email is required") @Email(message = "Invalid email format") String email,
			@NotBlank(message = "Password is required") String password) {
		super(email, password);
	}
	
}