# src/main/java/genaidemopoc/ecommerceproj1a/jwtspringsecurity/usersvc/model/Address.java
package genaidemopoc.ecommerceproj1a.jwtspringsecurity.usersvc.model;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Field;

public class Address {

	@Id
	private String id;

	@Field("street")
	private String street;

	@Field("city")
	private String city;

	@Field("state")
	private String state;

	@Field("zip_code")
	private String zipCode;

	// Default constructor
	public Address() {
	}

	// Parameterized constructor
	public Address(String street, String city, String state, String zipCode) {
		this.street = street;
		this.city = city;
		this.state = state;
		this.zipCode = zipCode;
	}

	// Getters and Setters

	public String getId() {
		return id;
	}

	public void setId(String id) {
		this.id = id;
	}

	public String getStreet() {
		return street;
	}

	public void setStreet(String street) {
		this.street = street;
	}

	public String getCity() {
		return city;
	}

	public void setCity(String city) {
		this.city = city;
	}

	public String getState() {
		return state;
	}

	public void setState(String state) {
		this.state = state;
	}

	public String getZipCode() {
		return zipCode;
	}

	public void setZipCode(String zipCode) {
		this.zipCode = zipCode;
	}
}